pyproject.toml
[tool.poetry]
name = "anomaly-detection"
version = "0.1.0"
description = "Using AI to detect anomalies in images from scavenge port inspections"
authors = ["Christoffer Sommerlund <christoffer.sommerlund@man-es.com>"]
readme = "README.md"
packages = [
    { include = "patchcore", from = "src/patchcore-inspection/src" },
    { include = "SupContrast", from = "src" },
    { include = "anomaly_detection", from = "src" },
    { include = "rd4ad", from = "src" },
    { include = "dimo", from = "src" },
]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
torch = "2.2.0+cu121"
torchvision = "0.17.0+cu121"
numpy = "^1.26.4"
pandas = "^2.2.0"
scipy = "^1.12.0"
scikit-learn = "^1.4.0"
scikit-image = "^0.22.0"
pillow = "^10.2.0"
matplotlib = "^3.8.2"
tqdm = "^4.66.1"
boto3 = "^1.34.37"
opencv-python = "^4.9.0.80"
optuna = "^3.5.0"
optuna-dashboard = "^0.14.0"
difpy = "^4.0.1"
boto3-stubs = { version = "^1.34.42", extras = ["essential"] }
torchmetrics = "^1.3.1"
timm = "^0.9.16"
click = "^8.1.7"
faiss-gpu = {version="^1.7.2", markers = "sys_platform != 'win32'"}
pretrainedmodels = "^0.7.4"
holoviews = "^1.18.2"
hvplot = "^0.9.2"
plotille = "^5.0.0"
typer = "^0.9.0"
ruff = "^0.2.0"
pre-commit = "^3.6.0"
pytest = "^8.0.0"
poetry = "^1.8.2"
ipywidgets = "^8.1.1"
ipykernel = "^6.29.0"
jupyter = "^1.0.0"
jupyter-bokeh = "^3.0.7"
trimesh = "^4.2.0"
pyrender = "^0.1.45"
scalene = "^1.5.37"
pynvml = "^11.5.0"
selenium = "^4.19.0"
datashader = "^0.16.1"
segment-anything = {git = "https://github.com/facebookresearch/segment-anything.git"}

[tool.ruff]
line-length = 88 # NOTE: This is the default value, may be subject to change!

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "cuda"
url = "https://download.pytorch.org/whl/cu121"
priority = "supplemental"

[[tool.poetry.source]]
name = "artifactory.ceon-dev"
url = "https://artifactory.ceon-dev.io/artifactory/api/pypi/pypi-team-analytics/simple/"
priority = "primary"
